organize input into Service package
